Add book methods to check read status
In this commit, two things were done. The first one was including
methods to check whether a book was read and to toggle the read
status when the Read button is clicked for each book on the bookshelf.

The second change was to further modularize displayBooks() and
displayBookDialog(). More work needs to be done to break down these
functions into more maintainable components. The Delete button was
added and fixed the read/unread button not updating the isRead value
of each book. Previously, each isRead value was being added as a starting
containing the true/false values as a string rather than a boolean value.
This was fixed in order for the toggleIsRead() function to do its job
the right way.

In the next commits, I am going to fix the displayBooks() function.
Currently, it re-renders the entire book list. It should only re-render
newer additions to it. I am also going to implement the behavior of
the delete button, namely delete the current book from the bookList.