Remove Show All Books button and hotfix
The previous functionality of the application included a Show All Books
button. This button was responsible for displaying all books once the
user desired to see them all in a list. Implementation wise, I was
dealing with duplicate books being added to the bookshelf HTMLElement.
This issue could've been fixed. However, as I thought about its
functionality and the user experience, I realized that it is more
convenient to display each book once it is added rather than hiding
all the books and displaying them all at once upon request. This makes
the library more interactive and easier to maintain.

In the next commit(s), I plan to fix the radio buttons on the
AddBookDialog element. Currently, there is a bug where all the items
are being added as read. The final phase of the project will be
updating the book stats that are below the Add Book button.