Fix Radio Buttons Not Marking Books as Read/Unread
In this commit, I fixed the value the radio buttons were sending to
the dialog form event handler. It turns out that whenever new Books
were added to the library, the button event handler was creating
duplicate dialogs without destroying the previous ones. This caused
multiple instances of the radio buttons, which all shared the same
ids and were conflicting with each other.

I added two calls to removeChild in order to remove the dialogs from
the document flow once they performed their tasks. This avoids conflicts
between the radio buttons and renders the book with the right read value.